requirements.txt
flask
flask_sqlalchemy
pandas
matplotlib
requests
scikit-learn
Pillow
colorthief
llama-cpp-python
transformers
datetime